Consolidate form render observer API
