+++ extension/ts/constants.ts
@@ -1,4 +1,3 @@
-const FORM_RENDER_EVENT = "library-thing-form-rendered";
 const FORM_DATA_ELEMENT_TAGS = ["textarea", "input", "select"];
 
-export {FORM_RENDER_EVENT, FORM_DATA_ELEMENT_TAGS};
+export {FORM_DATA_ELEMENT_TAGS};
