+++ extension/ts/extensions/pdf.ts
@@ -1,8 +1,8 @@
 import {invokeWorker} from "../workers/invoker";
 import {WorkerKind} from "../workers/types";
-import {FORM_RENDER_EVENT} from "../constants";
 import {showToast, ToastType} from "../ui/toast";
 import {createLoader, removeLoader} from "../ui/loadingIndicator";
+import {FORM_RENDER_EVENT} from "../services/renderFormObserver";
 
 const findTextContent = (id: string) => (): string =>
 	(document.getElementById(id) as HTMLTextAreaElement | HTMLInputElement)?.value ?? "";
